refactor: move disclaimer text component
