+++ src/app/export-modal/export-modal.component.spec.ts
@@ -2,11 +2,11 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { ExportModal } from './export-modal.component';
 import { CarbonEstimation } from '../types/carbon-estimator';
 import { FormsModule } from '@angular/forms';
-import { DisclaimerTextComponent } from '../disclaimer-text/disclaimer-text.component';
 import { InputGroupDisplay } from '../input-group-display/input-group-display.component';
 import { CarbonEstimationTreemapComponent } from '../carbon-estimation-treemap/carbon-estimation-treemap.component';
 import { CarbonEstimationTableComponent } from '../carbon-estimation-table/carbon-estimation-table.component';
 import { inputBinding } from '@angular/core';
+import { DisclaimerTextComponent } from '../components/disclaimer-text/disclaimer-text.component';
 
 describe('ExportModal', () => {
   let component: ExportModal;
