+++ src/app/export-modal/export-modal.component.ts
@@ -4,8 +4,8 @@ import { CarbonEstimationTableComponent } from '../carbon-estimation-table/carbo
 import { CarbonEstimation, EstimatorValues } from '../types/carbon-estimator';
 import { FormsModule } from '@angular/forms';
 import { InputGroupDisplay } from '../input-group-display/input-group-display.component';
-import { DisclaimerTextComponent } from '../disclaimer-text/disclaimer-text.component';
 import { A11yModule } from '@angular/cdk/a11y';
+import { DisclaimerTextComponent } from '../components/disclaimer-text/disclaimer-text.component';
 
 @Component({
   selector: 'export-modal',
